Use TreatmentRegimen in ReusableNotification
Use TreatmentRegimen in the format found in NTBS in the
ReusableNotification table, instead of the ShortCourse/MDRTreatment
style as found in ETS. Make the same changes in
ReusableNotification_ETS.

Add uspGenerateReusableNotificationTreatmentRegimen_ETS to convert
ETS-style treatment regimen in NTBS-style, to go into
ReusableNotification_ETS and then ReusableNotification.

Update uspMoveRecordsToLegacyExtract to not get the now-removed
ShortCourse/MDRTreatment fields from ReusableNotification, populate
instead with blank placeholders.

Move the uspGenerateReusableNotificationTreatmentRegimen procedure to
uspUpdateLegacyTreatmentRegimen: it converts NTBS-style
TreatmentRegimen to ETS-style. Use this to add treatment regimens for
NTBS notifications to the LegacyExtract.

Remove uspReusableNotification now that the ReusableNotification schema
has changed, it was used in now unsupported SSRS reports. Similarly,
remove references to ShortCourse/MDRTreatment in
ufnAuthorizedReusableNotification, another function used for deprecated
SSRS reports.