Keep Record_PersonalDetails.Postcode when invalid
When populating Record_PersonalDetails.Postcode, first set this to the
NTBS Patients.Postcode. This is what the user entered, and may be
invalid, and may not be formatted correctly.

Replace Record_PersonalDetails.Postcode with a formatted version of
Patients.PostcodeToLookup, if it exists and is valid. If it doesn't
exist or is invalid, leave Record_PersonalDetails.Postcode as what the
user originally entered.