Use specific HIV test values in reports
Use the more specific 'Offered' values (for 'and done', 'but not done'
and 'but refused') in Record_CaseData. Use the old ufnGetHivValue
stored proc to map from the NTBS values to these (basically adding
spaces). ETS has the values as we want them, so remove the mapping.

Remove the now unneeded mapping functions ufnGetHivTestOffered and
ufnGetHivTestOffered_ETS.

Update the KPI views to account for the different types of 'Offered'
values.